// src/notification/enums/notification-type.enum.ts
export enum NotificationType {
  FOLLOW = 'FOLLOW',
  POST_VOTE = 'POST_VOTE',
  QUESTION_VOTE = 'QUESTION_VOTE',
  COMMENT_ON_POST = 'COMMENT_ON_POST',
  COMMENT_VOTE = 'COMMENT_VOTE',
  ANSWER_ON_QUESTION = 'ANSWER_ON_QUESTION',
  ANSWER_VOTE = 'ANSWER_VOTE',
  ANSWER_ACCEPTED = 'ANSWER_ACCEPTED',
}